Added more weapon and bullet prefabs; added simple weapon selection
