+++ unity/Assets/Prefabs/Characters/Player.prefab
@@ -33,7 +33,9 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 3993694487903876632}
+  - {fileID: 9017256971741341929}
+  - {fileID: 1660227643658377513}
+  - {fileID: 1989377768490657356}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -163,7 +165,131 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!1001 &8564677815166003117
+--- !u!1001 &4365133261641349468
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 2039969185222439995}
+    m_Modifications:
+    - target: {fileID: 2890040985211317996, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_Name
+      value: BasicWeapon
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+--- !u!4 &9017256971741341929 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 4735441422055985589, guid: 298599c3b8ce1d74f99a255a47d77f65, type: 3}
+  m_PrefabInstance: {fileID: 4365133261641349468}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &6250427683526396060
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 2039969185222439995}
+    m_Modifications:
+    - target: {fileID: 2890040985211317996, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_Name
+      value: HeavyWeapon
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+--- !u!4 &1660227643658377513 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 4735441422055985589, guid: c23c4cf927ec1b846a23d96a6e16ffa6, type: 3}
+  m_PrefabInstance: {fileID: 6250427683526396060}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &6497597396478900217
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
@@ -172,11 +298,11 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 2890040985211317996, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
       propertyPath: m_Name
-      value: BasicWeapon
+      value: AssaultWeapon
       objectReference: {fileID: 0}
     - target: {fileID: 4735441422055985589, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
       propertyPath: m_RootOrder
-      value: 0
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 4735441422055985589, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
       propertyPath: m_LocalPosition.x
@@ -184,7 +310,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4735441422055985589, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
       propertyPath: m_LocalPosition.y
-      value: 0.2
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 4735441422055985589, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
       propertyPath: m_LocalPosition.z
@@ -220,8 +346,8 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
---- !u!4 &3993694487903876632 stripped
+--- !u!4 &1989377768490657356 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4735441422055985589, guid: a2587812d502a714ab2b1292aa9446dd, type: 3}
-  m_PrefabInstance: {fileID: 8564677815166003117}
+  m_PrefabInstance: {fileID: 6497597396478900217}
   m_PrefabAsset: {fileID: 0}
