+++ unity/Assets/Prefabs/Projectile/JoyProjectile.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 298599c3b8ce1d74f99a255a47d77f65
+guid: 1c32b1cbd49060c4db5c8b403d88b10f
 PrefabImporter:
   externalObjects: {}
   userData:
