+++ unity/Assets/Prefabs/Projectile/SmallLikeProjectile.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 3077807486098178287}
   - component: {fileID: 6235459332275871271}
   m_Layer: 0
-  m_Name: LikeProjectile
+  m_Name: SmallLikeProjectile
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -29,7 +29,7 @@ Transform:
   m_GameObject: {fileID: 8329337649942517644}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 1}
+  m_LocalScale: {x: 0.25, y: 0.25, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -77,7 +77,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_Sprite: {fileID: -2413806693520163455, guid: a86470a33a6bf42c4b3595704624658b, type: 3}
-  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_Color: {r: 1, g: 0.2688679, b: 0.4641898, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
@@ -99,8 +99,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d1106edf6dd9f834a91733683b56e0ca, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  movementSpeed: 10
-  dealtDamage: 4
+  movementSpeed: 15
+  dealtDamage: 1
   movementDirection: {x: 0, y: -1}
 --- !u!50 &3077807486098178287
 Rigidbody2D:
