+++ unity/Assets/Prefabs/Projectile/SmallLikeProjectile.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 298599c3b8ce1d74f99a255a47d77f65
+guid: 522a99662bed67b41806e8aacf07e2f9
 PrefabImporter:
   externalObjects: {}
   userData:
