+++ unity/Assets/Prefabs/Weapons/AssaultWeapon.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 298599c3b8ce1d74f99a255a47d77f65
+guid: a2587812d502a714ab2b1292aa9446dd
 PrefabImporter:
   externalObjects: {}
   userData:
