+++ unity/Assets/Prefabs/Weapons/HeavyWeapon.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 298599c3b8ce1d74f99a255a47d77f65
+guid: c23c4cf927ec1b846a23d96a6e16ffa6
 PrefabImporter:
   externalObjects: {}
   userData:
